ls-remote: Remove conflicting option "-h"
The "-h" option is documented to limit the result to heads, which is a
bit nonintuitive since "-h" is commonly used as the short option for
"--help". However, "-h" currently displays the usage information
which proves that there's a double reference to the option.

This patch will remove "-h" as a short option for "--heads". This
will maintain the current behaviour and correct the misleading
documentation.

Signed-off-by: Mattias Hansson <hansson.mattias@gmail.com>